requirements.txt
asgiref==3.2.10
Django==3.0.8
django-extensions==3.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
